Added Random Graph Gen with exponential decay probability based on edge length Changed everything to doubles Started Implementing BallTree ADT for use in KNN graph generation

+++ NetMining/ADT/IndexedItem.cs
@@ -3,10 +3,10 @@
     public class IndexedItem
     {
         public int HeapIndex;
-        public float NodeWeight;
+        public double NodeWeight;
         public int NodeIndex;
 
-        public IndexedItem(int nodeIndex, float nodeWeight)
+        public IndexedItem(int nodeIndex, double nodeWeight)
         {
             NodeIndex = nodeIndex;
             HeapIndex = -1;
